skills/1m-trade/skills/1m-trade-dex/scripts/requirements.txt
hyperliquid-python-sdk
websockets>=11.0,<12.0
python-dotenv